fix: fix end slash and types/swagger definition
