+++ pkg/api/favorite.go
@@ -1,68 +1,41 @@
 package api
 
-import "context"
+import (
+	"context"
+	"github.com/PlanVX/aweme/pkg/types"
+)
 
-type FavoriteActionReq struct {
-	Token      string `query:"token" json:"token"`             // 用户鉴权token
-	VideoID    int64  `query:"video_id" json:"video_id"`       // 视频id
-	ActionType int32  `query:"action_type" json:"action_type"` // 1-点赞，2-取消点赞
-}
-
-type FavoriteActionResp struct {
-	StatusCode int32  `json:"status_code"`          // 状态码，0-成功，其他值-失败
-	StatusMsg  string `json:"status_msg,omitempty"` // 返回状态描述
-}
-
-type FavoriteActionApiParam struct{}
-
-// NewFavoriteActionApi godoc
+// NewFavoriteAction godoc
 // @Summary 赞操作
 // @Description 赞操作
-// @Tags 基础接口
-// @Accept x-www-form-urlencoded
+// @Tags 互动接口
 // @Produce json
-// @Param favorite query FavoriteActionReq true "用户视频信息"
-// @Success 200 {object} FavoriteActionResp
-// @Router /favorite/action [get]
-func NewFavoriteActionApi(param FavoriteActionApiParam) *Api {
+// @Param favorite query types.FavoriteActionReq true "用户消息信息"
+// @Success 200 {object} types.FavoriteActionResp
+// @Router /favorite/action/ [get]
+func NewFavoriteAction() *Api {
 	return &Api{
 		Method: "POST",
 		Path:   "/favorite/action/",
-		Handler: WrapperFunc(func(ctx context.Context, req *FavoriteActionReq) (*FavoriteActionResp, error) {
+		Handler: WrapperFunc(func(ctx context.Context, req *types.FavoriteActionReq) (*types.FavoriteActionResp, error) {
 			return nil, nil
 		}),
 	}
 }
 
-// 用户点赞列表请求
-type FavoriteListReq struct {
-	UserId int64  `json:"user_id"` // 用户id
-	Token  string `json:"token"`   // 用户鉴权token
-}
-
-// 用户点赞列表响应
-type FavoriteListResp struct {
-	StatusCode int32   `json:"status_code"` // 状态码，0-成功，其他值-失败
-	StatusMsg  string  `json:"status_msg"`  // 返回状态描述
-	VideoList  []Video `json:"video_list"`  // 用户点赞视频列表
-}
-
-type NewFavoriteListApiParam struct{}
-
-// NewFavoriteListApi godoc
+// NewFavoriteList godoc
 // @Summary 收藏列表
 // @Description 收藏列表
-// @Tags 基础接口
-// @Accept x-www-form-urlencoded
+// @Tags 互动接口
 // @Produce json
-// @Param favorite query string true "用户信息"
-// @Success 200 {object} FavoriteListResp
-// @Router /favorite/list [get]
-func NewFavoriteListApi(param NewFavoriteListApiParam) *Api {
+// @Param favorite query types.FavoriteListReq true "请求信息"
+// @Success 200 {object} types.FavoriteListResp
+// @Router /favorite/list/ [get]
+func NewFavoriteList() *Api {
 	return &Api{
 		Method: "GET",
 		Path:   "/favorite/list/",
-		Handler: WrapperFunc(func(ctx context.Context, req *FavoriteListReq) (*FavoriteListResp, error) {
+		Handler: WrapperFunc(func(ctx context.Context, req *types.FavoriteListReq) (*types.FavoriteListResp, error) {
 			return nil, nil
 		}),
 	}
