+++ pkg/api/message.go
@@ -1,67 +1,43 @@
 package api
 
-import "context"
+import (
+	"context"
+	"github.com/PlanVX/aweme/pkg/types"
+)
 
-type MessageChatReq struct {
-	Token    string `json:"token" query:"token"`           // 用户鉴权token
-	ToUserID int64  `json:"to_user_id" query:"to_user_id"` // 对方用户id
-}
-
-type MessageChatResp struct {
-	StatusCode  int32     `json:"status_code"`  // 状态码，0-成功，其他值-失败
-	StatusMsg   string    `json:"status_msg"`   // 返回状态描述
-	MessageList []Message `json:"message_list"` // 消息列表
-}
-
-type MessageActionReq struct {
-	Token      string `json:"token" query:"token"`             // 用户鉴权token
-	ToUserID   int64  `json:"to_user_id" query:"to_user_id"`   // 对方用户id
-	ActionType int32  `json:"action_type" query:"action_type"` // 1-发送消息
-	Content    string `json:"content" query:"content"`         // 消息内容
-}
-
-type MessageActionResp struct {
-	StatusCode int32  `json:"status_code"` // 状态码，0-成功，其他值-失败
-	StatusMsg  string `json:"status_msg"`  // 返回状态描述
-}
-
-type MessageChatApiParam struct{}
-
-// NewMessageChatApi godoc
+// NewMessageChat godoc
 // @Summary 聊天记录
 // @Description 聊天记录
-// @Tags 基础接口
+// @Tags 社交接口
 // @Accept x-www-form-urlencoded
 // @Produce json
-// @Param message query MessageChatReq true "用户信息"
-// @Success 200 {object} MessageChatResp
-// @Router /message/chat [get]
-func NewMessageChatApi(param MessageChatApiParam) *Api {
+// @Param message query types.MessageListReq true "用户信息"
+// @Success 200 {object} types.MessageListResp
+// @Router /message/chat/ [get]
+func NewMessageChat() *Api {
 	return &Api{
 		Method: "GET",
 		Path:   "/message/chat/",
-		Handler: WrapperFunc(func(ctx context.Context, req *MessageChatReq) (*MessageChatResp, error) {
+		Handler: WrapperFunc(func(ctx context.Context, req *types.MessageListReq) (*types.MessageListResp, error) {
 			return nil, nil
 		}),
 	}
 }
 
-type MessageActionApiParam struct{}
-
-// NewMessageActionApi godoc
+// NewMessageAction godoc
 // @Summary 消息操作
 // @Description 消息操作
-// @Tags 基础接口
+// @Tags 社交接口
 // @Accept x-www-form-urlencoded
 // @Produce json
-// @Param message query MessageActionReq true "用户消息信息"
-// @Success 200 {object} MessageActionResp
-// @Router /message/action [post]
-func NewMessageActionApi(param MessageActionApiParam) *Api {
+// @Param message query types.MessageActionReq true "用户消息"
+// @Success 200 {object} types.MessageActionResp
+// @Router /message/action/ [post]
+func NewMessageAction() *Api {
 	return &Api{
 		Method: "POST",
 		Path:   "/message/action/",
-		Handler: WrapperFunc(func(ctx context.Context, req *MessageActionReq) (*MessageActionResp, error) {
+		Handler: WrapperFunc(func(ctx context.Context, req *types.MessageActionReq) (*types.MessageActionResp, error) {
 			return nil, nil
 		}),
 	}
